refactor: renamed TestContainer to TestSuite.
