+++ src/Beta/BetaTest.cs
@@ -5,12 +5,12 @@ namespace Beta;
 /// <summary>
 ///     Defines a single test or test suite.
 /// </summary>
-/// <param name="Container">The instance of the type holding the test.</param>
+/// <param name="Suite">The instance of the type holding the test.</param>
 /// <param name="Input">The optional input for the test.</param>
 /// <param name="TestName">The name of the test.</param>
 /// <param name="Apply">The method, that when executed will apply input to the test and create a proof.</param>
 [PublicAPI]
-public record BetaTest(object Container, IEnumerable<object>? Input, string TestName, Func<object, Proof> Apply)
+public record BetaTest(object Suite, IEnumerable<object>? Input, string TestName, Func<object, Proof> Apply)
 {
     /// <summary>
     ///     Gets the traits for the test.
@@ -20,11 +20,11 @@ public record BetaTest(object Container, IEnumerable<object>? Input, string Test
     /// <summary>
     ///     Initializes a new instance of the <see cref="BetaTest" /> class.
     /// </summary>
-    /// <param name="container">The instance of the type holding the test.</param>
+    /// <param name="suite">The instance of the type holding the test.</param>
     /// <param name="testName">The name of the test.</param>
     /// <param name="apply">The apply method for the test.</param>
-    public BetaTest(object container, string testName, Func<Proof> apply)
-        : this(container, null, testName, _ => apply())
+    public BetaTest(object suite, string testName, Func<Proof> apply)
+        : this(suite, null, testName, _ => apply())
     {
     }
 
@@ -40,9 +40,9 @@ public record BetaTest(object Container, IEnumerable<object>? Input, string Test
         (Input ?? Array.Empty<object>()).ToDictionary(_ => Guid.NewGuid());
 
     /// <summary>
-    ///     Gets the FQN of the container for the test.
+    ///     Gets the FQN of the suite for the test.
     /// </summary>
-    public string FullyQualifiedTypeName { get; } = Container.GetType().FullName!;
+    public string FullyQualifiedTypeName { get; } = Suite.GetType().FullName!;
 
     /// <summary>
     ///     Gets or sets the skip reason for this test.
@@ -82,12 +82,12 @@ public record BetaTest(object Container, IEnumerable<object>? Input, string Test
     // /// <summary>
     // ///     Gets the assembly that contains the test.
     // /// </summary>
-    // public Assembly Assembly => Container.GetType().Assembly;
+    // public Assembly Assembly => Suite.GetType().Assembly;
     //
     // /// <summary>
     // ///     Gets the declaring type of the test.
     // /// </summary>
-    // public string DeclaringType => Container.GetType().FullName!;
+    // public string DeclaringType => Suite.GetType().FullName!;
     //
     // /// <summary>
     // ///     Gets the method name of the test.
@@ -105,12 +105,12 @@ public record BetaTest<TInput> : BetaTest
 {
     /// <summary>
     /// </summary>
-    /// <param name="container">The instance of the type holding the test.</param>
+    /// <param name="suite">The instance of the type holding the test.</param>
     /// <param name="input">The input for the test cases.</param>
     /// <param name="testName">The name of the test.</param>
     /// <param name="apply">The apply method for the test.</param>
-    public BetaTest(TestContainer container, IEnumerable<TInput> input, string testName, Func<TInput, Proof> apply)
-        : base(container, (IEnumerable<object>)input, testName, o => apply((TInput)o))
+    public BetaTest(object suite, IEnumerable<TInput> input, string testName, Func<TInput, Proof> apply)
+        : base(suite, (IEnumerable<object>)input, testName, o => apply((TInput)o))
     {
     }
 }
