+++ src/Beta/Internal/Execution/TestContainerProcessor.cs
@@ -10,7 +10,7 @@ public class TestContainerProcessor : ITestSuiteProcessor
 {
     /// <inheritdoc />
     public void PreProcess(object instance) =>
-        (instance as TestContainer)?.Initialize();
+        (instance as TestSuite.DI)?.Initialize();
 
     /// <inheritdoc />
     [ExcludeFromCodeCoverage]
